Initial commit
Finished implementing the Deposit Screen

Added internationalization capabilities to the DashboardScreen, DepositScreen and EarnScreen

Minor fixes for the splash screen in Android

Implemented the UI for the Transfer Screen

Implemented the network queries to transfer coins between accounts

Added no connectivity warning on the transfer screen

Fixes for the Transfer and Earn Screens

Auto selection of the destination plan when transfering from non current interest plan accounts

Added a dialog to prevent users from trying to close interest plans with balance

Code quality improvements on the EarnScreen

Added a trailing text option the the amoun input field in the TransferScreen

Added a UI for the Withdraw screen

It still lacks all the logic to confirm fees and actually withdraw.
In the process tags and memos were added to the deposit screen as well,
and the warning banner was centralized.

Migration from Double to BigNum

This will avois problems with decimal inaccuracy

Minor UI changes due to coindepo token turning into crypto

Remember the selected coin in the earn screen

Added an option to mask balances on the main screen

Added masking to the email address in the verification code input

Added logic to allow re-sending the email verification code

Refactored the LoggedInUseCase class to make it easier to use

Added confirmation functionallity for the withdraw screen

Complete implementation of the Withdraw functionality

Minor fixes for the resend email verification code functionality.

Added transactions screen functionality

Create LICENSE

Added copyright to all files in the project

Modified readme

Proper readme